Add a userscript to suppress repetitive text
This is the initial version of the suppress_repetitions userscript.
The script examines the context and identifies tokens that would
cause additional "repetitious" text in the story.  It adds a negative
bias to these and similar tokens in order to reduce the chance
of the story generator repeating itself.  The penalty starts small
but increases the longer (and the more frequent) the repetitious text.

There are several known limitations in this version including an
increased likelihood of pronoun misattribution due to suppressing
correct pronouns that could cause repetition. There is also not
currently a meaningful configuration file associated with this script;
to edit the parameters of the bias function you will need to edit
the Lua script itself.  These deficiencies will be improved in a
future version.